fix(Reward): list and budget
